remove MateTaskAssignmentDataSource interface and update tests
