--- src/main/kotlin/data/source/remote/mongo/MateTaskAssignmentMongoDataSource.kt
@@ -1,36 +1,0 @@
-package data.source.remote.mongo
-
-import com.mongodb.kotlin.client.coroutine.MongoCollection
-import data.dto.MateTaskAssignmentDto
-
-import kotlinx.coroutines.flow.toList
-import org.example.data.source.MateTaskAssignmentDataSource
-import org.litote.kmongo.and
-import org.litote.kmongo.eq
-
-class MateTaskAssignmentMongoDataSource(
-    private val mateTaskAssignmentCollection: MongoCollection<MateTaskAssignmentDto>,
-) : MateTaskAssignmentDataSource {
-
-    override suspend fun addUserToTask(username: String, taskId: String): Boolean =
-        mateTaskAssignmentCollection.insertOne(
-            MateTaskAssignmentDto(
-                username = username,
-                taskId = taskId
-            )
-        ).insertedId != null
-
-    override suspend fun deleteUserFromTask(username: String, taskId: String): Boolean =
-        mateTaskAssignmentCollection.deleteOne(
-            and(
-                MateTaskAssignmentDto::username eq username,
-                MateTaskAssignmentDto::taskId eq taskId
-            )
-        ).deletedCount > 0
-
-    override suspend fun getUsersMateTaskByTaskId(taskId: String): List<MateTaskAssignmentDto> =
-        mateTaskAssignmentCollection.find(MateTaskAssignmentDto::taskId eq taskId).toList()
-
-    override suspend fun getUsersMateTaskByUserName(username: String): List<MateTaskAssignmentDto> =
-        mateTaskAssignmentCollection.find(MateTaskAssignmentDto::username eq username).toList()
-}
